pyproject.toml
[project]
name = "leak-response-scada"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "attrs>=25.3.0",
    "cattrs>=25.2.0",
    "coolprop>=7.0.0",
    "nicegui>=2.22.2",
    "orjson>=3.11.2",
    "pint>=0.25",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "scipy>=1.16.2",
]
